Fix: RestaurantRepositoryImpl의 카운트 세는 방식 변경 및 정확도순 정렬 추가 및 테스트 코드 추가

+++ src/main/java/com/posomo/saltit/domain/restaurant/dto/RestaurantSummary.java
@@ -33,7 +33,7 @@ public class RestaurantSummary {
 	private String categoryName;
 
 	@Schema(description = "가격 내에 메뉴 갯수", example = "5")
-	private Long menuSize;
+	private Integer menuSize;
 
 	@Schema(description = "경도 값", example = "127.0502")
 	private Double longitude;
